Do not count ship moving slower than 2kts in this feature (AISAB-38)

+++ ais-ab-stat-builder/src/main/java/dk/dma/ais/abnormal/stat/features/ShipTypeAndSizeFeature.java
@@ -72,7 +72,7 @@ public class ShipTypeAndSizeFeature implements Feature {
 
         Track track = event.getTrack();
 
-        Double sog = (Double) track.getProperty(Track.SPEED_OVER_GROUND);
+        Float sog = (Float) track.getProperty(Track.SPEED_OVER_GROUND);
         if (sog != null && sog < 2.0) {
             return; // If track has a sog and it is < 2 - don't run this feature
         }
